requirements.txt
fastapi
uvicorn
httpx
pydantic<2
pytest
pytest-asyncio
sqlalchemy
psycopg2-binary
xmltodict
python-dateutil
python-multipart
